refactor SAST Transformer
